Make feature "Set image as a wallpaper" optional
This single feature depends on the `ashpd` crate which is a pretty huge
dependency - it adds 61 transitive dependencies and doubles the size of
the image-roll binary. Moreover, this feature is usable only on some
desktops. For example, Sway users will not benefit from it.

This commit introduces a (cargo) feature named "wallpaper", which is
enabled by default. When the application is built with
`--no-default-features`, the `ashpd` dependency is excluded from the
dependency tree and the "Set as wallpaper" button is disabled (by
setting `sensitive` to `false`).

This is suboptiomal, ideally I'd like to remove the button from the UI
when this feature is disabled, but I don't know how to do it without
adding too much complexity.